folder restructure to match structure.json
